+++ mrblib/acli/cli.rb
@@ -16,7 +16,7 @@ module Acli
       uri.instance_variable_set(:@path, DEFAULT_CABLE_PATH) if uri.path.nil? || uri.path.empty?
 
       poller = Poller.new
-      socket = Socket.new(uri, @options.delete(:headers))
+      socket = Socket.new(uri, @options.delete(:headers), @options.delete(:protocol))
       client = Client.new(Utils.uri_to_ws_s(uri), socket, **@options)
 
       poller.add_client(client)
@@ -47,6 +47,8 @@ Options:
                         # Example: `--headers="x-api-token:secret,cookie:user_id=26"`
                         # NOTE: the value should not contain whitespaces.
 
+  --sub-protocol        # Custom WebSocket subprotocol
+
   --quit-after          # Automatically quit after an even occured.
                         # Possible values are:
                         #  - connected — quit right after successful connection ("welcome" message)
@@ -65,7 +67,15 @@ Options:
 
     def parse_options(argv)
       class << argv; include Getopts; end
-      argv.getopts("vhu:c:", "url:", "channel:", "channel-params:", "headers:", "quit-after:").yield_self do |opts|
+      argv.getopts(
+        "vhu:c:",
+        "url:",
+        "channel:",
+        "channel-params:",
+        "headers:",
+        "sub-protocol:",
+        "quit-after:"
+      ).yield_self do |opts|
         print_version if opts["v"]
         print_help if opts["h"]
 
@@ -84,7 +94,8 @@ Options:
           url: opts["u"] || opts["url"],
           channel: channel,
           quit_after: opts["quit-after"],
-          headers: headers
+          headers: headers,
+          protocol: opts["sub-protocol"]
         }.tap { |data| data.delete_if { _2.empty? } }
       end
     end