+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Add WS sub-protocols support. ([@palkan][])
+
 - Upgrade to mruby 3.0. ([@palkan][])
 
 ## 0.3.1 (2020-06-30)